Fix 4 fractional share bugs found in deep audit
budget.py:
- Kelly overlay: int() → round(,6) to preserve fractional qty

trade_logger.py:
- log_trade_open qty: int → float

iceberg.py:
- should_use_iceberg, iceberg_buy, _plan_slices: int → float signatures

database.py:
- qty column: INTEGER → REAL in trade_log, shadow_trades, slippage_log
- Migration: CAST existing qty values to REAL on startup

+++ trade_logger.py
@@ -43,7 +43,7 @@ def log_trade_open(
     payload:      WebhookPayload,
     sentiment:    SentimentResult | None,
     order_result: dict,
-    qty:          int,
+    qty:          float,
     sizing_meta:  dict | None = None,
     slippage_meta: dict | None = None,
 ) -> int:
